Fix WebSocket duplication in handleServerMessagesGuard
Before, a new WebSocket was created on every render of
`handleServerMessagesGuard`, and event listeners bound to all of them.
This results in lots of sockets being opened - often about 8 on
initial page load, and then another for each page navigation.

This changes the hook to keep the WebSocket instance in a React ref,
so that the same socket instance is used consistently for the lifetime
of the wrapped component. The startup procedure is also tweaked to
unbind the startup message handlers and replace them with the
long-term message handler.