Delete outer session key
Since the outer session key is stored in IndexedDB under a constant name that is
shared between all tabs, concurrent tabs overwrite the existing outer session
key when the user re-authenticates in the new tab. This causes all attempts to
access encrypted data to fail because the inner session key can no longer be
unwrapped.

As noted in the self-critique [1] of the encryption architecture, the outer
session key doesn't really add much. Any malicious code with access to the
session storage probably also has access to the IndexedDB, and thus also to the
outer session key. We don't really gain anything by having the outer session key
unextractable.

[1]: https://wwwallet.github.io/wallet-docs/docs/wallet/encryption-architecture/#unextractable-cryptokeys